Group package-related functions
