+++ lib/deppack/deppack.js
@@ -9,7 +9,7 @@ const xBrowserResolve = require('./resolve');
 const loadGlobalsAndStyles = require('./load-globs-styles');
 const explore = require('./explore');
 const processFiles = require('./process');
-const helpers = require('./helpers');
+const packages = require('./packages');
 const DepGraph = require('./dep-graph');
 const UsedShims = require('./used-shims');
 const ProjectModuleMap = require('./project-module-map');
@@ -60,7 +60,7 @@ class Deppack {
     this.resolve = xBrowserResolve.resolve(this.rootPackage, this.modMap, this.npm.aliases, this.globalPseudofile);
 
     this.getPackageJson = filePath => {
-      return helpers.packageJson(filePath, this.rootPackage);
+      return packages.packageJson(filePath, this.rootPackage);
     };
   }
