+++ lib/deppack/explore.js
@@ -5,32 +5,15 @@ const promisify = require('../helpers').promisify;
 const each = promisify(require('async-each'));
 const detective = require('detective');
 
-const moduleNaming = require('./module-naming');
 const shims = require('./shims');
 const helpers = require('./helpers');
+const packages = require('./packages');
 
 const not = helpers.not;
 const isRelative = helpers.isRelative;
 const makeRelative = helpers.makeRelative;
 
 const isJs = (fileList, path) => fileList.find(path).type === 'javascript';
-
-const shouldIncludeDep = (getPackageJson, path) => {
-  const currRoot = moduleNaming.getModuleRootName(path);
-  const pkg = getPackageJson(sysPath.resolve(path));
-  const deps = Object.assign({}, pkg.dependencies || {}, pkg.peerDependencies || {});
-  return dep => {
-    const root = dep.split('/')[0];
-    const browser = typeof pkg.browser === 'object' ? pkg.browser : typeof pkg.browserify === 'object' ? pkg.browserify : {};
-    // ignore optional dependencies
-    return (isRelative(dep) ||
-            root in deps ||
-            browser[root] && browser[root] in deps ||
-            root === currRoot ||
-            root in shims.fileShims ||
-            shims.emptyShims.indexOf(root) !== -1);
-  };
-};
 
 const envify = (source, map) => {
   if (source.indexOf('process.env.') !== -1) {
@@ -64,7 +47,7 @@ const depsFor = (fileReflection, getPackageJson) => {
   return (path, source) => {
     const allDeps = isVendor(path) ? [] : detective(source).filter(x => x && x.length > 0);
     const usesProcess = isVendor(path) ? false : shims.shouldIncludeProcess(source);
-    const deps = isApp(path) ? allDeps.filter(x => !isRelative(x)) : isPackage(path) ? allDeps.filter(shouldIncludeDep(getPackageJson, path)) : allDeps;
+    const deps = isApp(path) ? allDeps.filter(x => !isRelative(x)) : isPackage(path) ? allDeps.filter(packages.canCallDep(getPackageJson, path)) : allDeps;
     if (usesProcess) deps.push('process');
     return deps;
   };
