+++ lib/deppack/helpers.js
@@ -1,9 +1,6 @@
 'use strict';
-
 const sysPath = require('path');
 const fs = require('fs');
-const deepExtend = require('../helpers').deepExtend;
-const getModuleRootPath = require('./module-naming').getModuleRootPath;
 
 const fsstat = require('../helpers').promisify(fs.stat);
 const isDir = path => fsstat(path).then(stat => stat.isDirectory(), () => Promise.resolve(false));
@@ -20,25 +17,4 @@ const uniq = list => {
   }, {}));
 };
 
-const applyPackageOverrides = (pkg, rootPackage) => {
-  const pkgOverride = rootPackage.overrides[pkg.name];
-
-  if (pkgOverride) {
-    pkg = deepExtend(pkg, pkgOverride);
-  }
-
-  return pkg;
-};
-
-const getDepPackageJson = (depPath, rootPackage) => {
-  const depJson = require(sysPath.join(depPath, 'package.json'));
-  applyPackageOverrides(depJson, rootPackage);
-  return depJson;
-};
-
-const packageJson = (filePath, rootPackage) => getDepPackageJson(
-  getModuleRootPath(filePath),
-  rootPackage
-);
-
-module.exports = {isDir, makeRelative, isRelative, not, uniq, applyPackageOverrides, packageJson};
+module.exports = {isDir, makeRelative, isRelative, not, uniq};
