+++ lib/deppack/resolve.js
@@ -5,7 +5,7 @@ const trueCasePath = require('true-case-path');
 const promisify = require('../helpers').promisify;
 const glob = promisify(require('glob'));
 const moduleNaming = require('./module-naming');
-const helpers = require('./helpers');
+const packages = require('./packages');
 const shims = require('./shims');
 
 const noPackage = rootPackage => x => rootPackage.dependencies[x] == null;
@@ -69,7 +69,7 @@ const checkImproperCase = (mod, opts, res) => {
 };
 
 const resolve = (rootPackage, modMap, aliases, globalPseudofile) => (mod, opts, cb) => {
-  const packageFilter = pkg => helpers.applyPackageOverrides(pkg, rootPackage);
+  const packageFilter = pkg => packages.applyPackageOverrides(pkg, rootPackage);
   const modules = modMap.toHash();
   Object.assign(opts, {packageFilter, modules, extensions: ['.js', '.json']});
   mod = aliases && aliases[mod] || mod;
